requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bizerror==0.4.11
certifi==2023.7.22
charset-normalizer==3.3.0
click==7.1.2
Django==4.2.6
django-bootstrap-v5==1.0.11
django-click==2.3.0
djangorestframework==3.14.0
docker==4.2.0
dockerpty==0.4.1
docopt==0.6.2
fastutils==0.43.1
gitdb==4.0.10
GitPython==3.1.0
idna==3.4
jsonformatter==0.3.2
msgpack==1.0.7
Pillow==10.1.0
popper==2.7.0
psutil==5.9.6
psycopg==3.1.14
psycopg-binary==3.1.14
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycryptodome==3.19.0
pykwalify==1.7.0
pypinyin==0.49.0
pypiwin32==223
python-box==4.2.3
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==5.3.1
requests==2.31.0
rsa==4.9
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
semver==3.0.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
spython==0.0.79
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
websocket-client==1.6.4
zenutils==0.5.1
